+++ quill/include/quill/Logger.h
@@ -168,7 +168,7 @@ public:
     write_buffer = detail::align_pointer<alignof(detail::Header), std::byte>(write_buffer);
 
     new (write_buffer) detail::Header(
-      get_metadata_ptr<TMacroMetadata, FmtArgs...>, std::addressof(_logger_details),
+      detail::get_metadata_and_format_fn<TMacroMetadata, FmtArgs...>, std::addressof(_logger_details),
       (_logger_details.timestamp_clock_type() == TimestampClockType::Rdtsc) ? quill::detail::rdtsc()
         : (_logger_details.timestamp_clock_type() == TimestampClockType::System)
         ? static_cast<uint64_t>(std::chrono::system_clock::now().time_since_epoch().count())